+++ src/app/components/category-picker/data/category-pager-data.interface.ts
@@ -1,0 +1,10 @@
+import { CategoryDataInterface } from './category-data.interface';
+import { TerraPagerInterface } from '../../../../';
+
+/**
+ * @author ziyad.hajj-hassan
+ */
+export interface CategoryPagerDataInterface extends TerraPagerInterface
+{
+    entries:Array<CategoryDataInterface>;
+}