+++ src/app/components/forms/multi-check-box/data/terra-multi-check-box-value.interface.ts
@@ -1,0 +1,6 @@
+export interface TerraMultiCheckBoxValueInterface
+{
+    value:any;
+    caption:string;
+    selected:boolean;
+}