+++ src/app/components/forms/multi-check-box/terra-multi-check-box.component.spec.ts
@@ -1,0 +1,11 @@
+// /* tslint:disable:no-unused-variable */
+//
+// import { TestBed, async } from '@angular/core/testing';
+// import { TerraMultiSelectBoxComponent } from './plenty-multi-select-box.component';
+//
+// describe('Component: MultiSelectBox', () => {
+//   it('should create an instance', () => {
+//     let component = new TerraMultiSelectBoxComponent();
+//     expect(component).toBeTruthy();
+//   });
+// });