+++ src/app/components/note/terra-note.component.ts
@@ -10,12 +10,15 @@ import {
 })
 export class TerraNoteComponent
 {
-    @Input() inputIsSelected:boolean;
-    @Input() inputId:number;
-    @Input() inputHeaderText:string;
-    @Input() inputNoteText:string;
+    @Input()
+    public inputIsSelected:boolean;
 
-    constructor()
-    {
-    }
+    @Input()
+    public inputId:number;
+
+    @Input()
+    public inputHeaderText:string;
+
+    @Input()
+    public inputNoteText:string;
 }